+++ ChangeLog.md
@@ -7,6 +7,7 @@ IMPROVEMENTS
 - switch to go 1.19
 - refactoring to remove legacy `storage` package
 - add `table` parameter to `tables` cli command and `/backup/tables` API handler, fix [367](https://github.com/AlexAkulov/clickhouse-backup/issues/367)
+- add `--resumable` parameter to `create_remote`, `upload`, `restore_remote`, `donwload` commands to allow resume upload or download after break. Ignored for `remote_storage: custom`, fix [207](https://github.com/AlexAkulov/clickhouse-backup/issues/207)
 
 BUG FIXES
 - fix wrong upload / download behavior for `compression_format: none` and `remote_storage: ftp`